+++ pkg/ingester/client/custom.go
@@ -5,27 +5,25 @@
 
 package client
 
-// ChunksCount returns the number of chunks in response.
-func (m *QueryStreamResponse) ChunksCount() int {
-	if len(m.Chunkseries) == 0 {
+func ChunksCount(series []TimeSeriesChunk) int {
+	if len(series) == 0 {
 		return 0
 	}
 
 	count := 0
-	for _, entry := range m.Chunkseries {
+	for _, entry := range series {
 		count += len(entry.Chunks)
 	}
 	return count
 }
 
-// ChunksSize returns the size of all chunks in the response.
-func (m *QueryStreamResponse) ChunksSize() int {
-	if len(m.Chunkseries) == 0 {
+func ChunksSize(series []TimeSeriesChunk) int {
+	if len(series) == 0 {
 		return 0
 	}
 
 	size := 0
-	for _, entry := range m.Chunkseries {
+	for _, entry := range series {
 		for _, chunk := range entry.Chunks {
 			size += chunk.Size()
 		}